Move download-related logic to DownloadManager [skip ci]
