+++ python/yugabyte_db_thirdparty/archive_handling.py
@@ -1,0 +1,35 @@
+# Copyright (c) Yugabyte, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+# in compliance with the License. You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software distributed under the License
+# is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+# or implied. See the License for the specific language governing permissions and limitations
+# under the License.
+
+from typing import Optional
+
+
+TAR_EXTRACT = 'tar --no-same-owner -xf {}'
+# -o -- force overwriting existing files
+ZIP_EXTRACT = 'unzip -q -o {}'
+
+ARCHIVE_TYPES = {
+    '.tar.bz2': TAR_EXTRACT,
+    '.tar.gz': TAR_EXTRACT,
+    '.tar.xz': TAR_EXTRACT,
+    '.tgz': TAR_EXTRACT,
+    '.zip': ZIP_EXTRACT,
+}
+
+
+def make_archive_name(name: str, version: str, download_url: Optional[str]) -> Optional[str]:
+    if download_url is None:
+        return '{}-{}{}'.format(name, version, '.tar.gz')
+    for ext in ARCHIVE_TYPES:
+        if download_url.endswith(ext):
+            return '{}-{}{}'.format(name, version, ext)
+    return None
