+++ yb_build_thirdparty_main.py
@@ -51,14 +51,13 @@ from yugabyte_db_thirdparty.custom_logging import (
 from yugabyte_db_thirdparty.string_util import indent_lines, split_into_word_set
 from yugabyte_db_thirdparty.util import (
     assert_list_contains,
-    hashsum_file,
-    compute_file_sha256,
     mkdir_if_missing,
     remove_path,
     PushDir,
     which_executable,
     which_must_exist,
     EnvVarContext,
+    YB_THIRDPARTY_DIR,
 )
 from yugabyte_db_thirdparty.os_detection import (
     is_mac,
@@ -66,16 +65,12 @@ from yugabyte_db_thirdparty.os_detection import (
     is_centos,
 )
 from yugabyte_db_thirdparty.multi_build import MultiBuilder
+from yugabyte_db_thirdparty.download_manager import DownloadManager
 import build_definitions
 from build_definitions import *  # noqa
 
 import_submodules(build_definitions)
 
-
-CLOUDFRONT_URL = 'http://d3dr9sfxru4sde.cloudfront.net/{}'
-MAX_FETCH_ATTEMPTS = 10
-INITIAL_DOWNLOAD_RETRY_SLEEP_TIME_SEC = 1.0
-DOWNLOAD_RETRY_SLEEP_INCREASE_SEC = 0.5
 
 PLACEHOLDER_RPATH = (
     "/tmp/making_sure_we_have_enough_room_to_set_rpath_later_{}_end_of_rpath".format('_' * 256))
@@ -155,6 +150,7 @@ class Builder(BuilderInterface):
     c_flags: List[str]
     cxx_flags: List[str]
     libs: List[str]
+    download_manager: DownloadManager
 
     """
     This class manages the overall process of building third-party dependencies, including the set
@@ -176,8 +172,6 @@ class Builder(BuilderInterface):
         self.linuxbrew_dir = None
         self.cc = None
         self.cxx = None
-
-        self.load_expected_checksums()
 
     def set_compiler(self, compiler_type: str) -> None:
         if is_mac():
@@ -212,6 +206,9 @@ class Builder(BuilderInterface):
         self.args = parse_cmd_line_args()
         if self.args.make_parallelism:
             os.environ['YB_MAKE_PARALLELISM'] = str(self.args.make_parallelism)
+        self.download_manager = DownloadManager(
+            should_add_checksum=self.args.add_checksum,
+            download_dir=self.tp_download_dir)
 
     def use_only_clang(self) -> bool:
         return is_mac() or self.args.single_compiler_type == 'clang'
@@ -337,7 +334,6 @@ class Builder(BuilderInterface):
         if self.args.clean:
             self.clean()
         self.prepare_out_dirs()
-        self.curl_path = which_must_exist('curl')
         os.environ['PATH'] = ':'.join([
                 os.path.join(self.tp_installed_common_dir, 'bin'),
                 os.path.join(self.tp_installed_llvm7_common_dir, 'bin'),
@@ -475,77 +471,12 @@ class Builder(BuilderInterface):
                     log("Removing %s source: %s", dependency.name, src_dir)
                     remove_path(src_dir)
 
-            archive_path = self.archive_path(dependency)
+            archive_path = self.get_archive_path(dependency)
             if archive_path is not None:
                 log("Removing %s archive: %s", dependency.name, archive_path)
                 remove_path(archive_path)
 
-    def download_dependency(self, dep: Dependency) -> None:
-        src_path = self.get_source_path(dep)
-        patch_level_path = os.path.join(src_path, 'patchlevel-{}'.format(dep.patch_version))
-        if os.path.exists(patch_level_path):
-            return
-
-        download_url = dep.download_url
-        if download_url is None:
-            download_url = CLOUDFRONT_URL.format(dep.get_archive_name())
-            log("Using legacy download URL: %s (we should consider moving this to GitHub)",
-                download_url)
-
-        archive_path = self.archive_path(dep)
-
-        remove_path(src_path)
-        # If download_url is "mkdir" then we just create empty directory with specified name.
-        if download_url != 'mkdir':
-            if archive_path is None:
-                return
-            self.ensure_file_downloaded(download_url, archive_path)
-            self.extract_archive(archive_path,
-                                 os.path.dirname(src_path),
-                                 os.path.basename(src_path))
-        else:
-            log("Creating %s", src_path)
-            mkdir_if_missing(src_path)
-
-        if hasattr(dep, 'extra_downloads'):
-            for extra in dep.extra_downloads:
-                assert extra.archive_name is not None
-                archive_path = os.path.join(self.tp_download_dir, extra.archive_name)
-                log("Downloading %s from %s", extra.archive_name, extra.download_url)
-                self.ensure_file_downloaded(extra.download_url, archive_path)
-                output_path = os.path.join(src_path, extra.dir_name)
-                self.extract_archive(archive_path, output_path)
-                if extra.post_exec is not None:
-                    with PushDir(output_path):
-                        assert isinstance(extra.post_exec, list)
-                        if isinstance(extra.post_exec[0], str):
-                            subprocess.check_call(cast(List[str], extra.post_exec))
-                        else:
-                            for command in extra.post_exec:
-                                subprocess.check_call(command)
-
-        if hasattr(dep, 'patches'):
-            with PushDir(src_path):
-                for patch in dep.patches:
-                    log("Applying patch: %s", patch)
-                    process = subprocess.Popen(['patch', '-p{}'.format(dep.patch_strip)],
-                                               stdin=subprocess.PIPE)
-                    with open(os.path.join(self.tp_dir, 'patches', patch), 'rt') as inp:
-                        patch = inp.read()
-                    assert process.stdin is not None
-                    process.stdin.write(patch.encode('utf-8'))
-                    process.stdin.close()
-                    exit_code = process.wait()
-                    if exit_code:
-                        fatal("Patch {} failed with code: {}".format(dep.name, exit_code))
-                if dep.post_patch:
-                    subprocess.check_call(dep.post_patch)
-
-        with open(patch_level_path, 'wb') as out:
-            # Just create an empty file.
-            pass
-
-    def archive_path(self, dep: Dependency) -> Optional[str]:
+    def get_archive_path(self, dep: Dependency) -> Optional[str]:
         archive_name = dep.get_archive_name()
         if archive_name is None:
             return None
@@ -553,169 +484,6 @@ class Builder(BuilderInterface):
 
     def get_source_path(self, dep: Dependency) -> str:
         return os.path.join(self.tp_src_dir, dep.get_source_dir_basename())
-
-    def get_checksum_file(self) -> str:
-        return os.path.join(self.tp_dir, CHECKSUM_FILE_NAME)
-
-    def load_expected_checksums(self) -> None:
-        checksum_file = self.get_checksum_file()
-        if not os.path.exists(checksum_file):
-            fatal("Expected checksum file not found at {}".format(checksum_file))
-
-        self.filename2checksum = {}
-        with open(checksum_file, 'rt') as inp:
-            for line in inp:
-                line = line.strip()
-                if not line or line.startswith('#'):
-                    continue
-                sum, fname = line.split(None, 1)
-                if not re.match('^[0-9a-f]{64}$', sum):
-                    fatal("Invalid checksum: '%s' for archive name: '%s' in %s. Expected to be a "
-                          "SHA-256 sum (64 hex characters).", sum, fname, checksum_file)
-                self.filename2checksum[fname] = sum
-
-    def get_expected_checksum(self, filename: str, downloaded_path: str) -> str:
-        if filename not in self.filename2checksum:
-            if self.args.add_checksum:
-                checksum_file = self.get_checksum_file()
-                with open(checksum_file, 'rt') as inp:
-                    lines = inp.readlines()
-                lines = [line.rstrip() for line in lines]
-                checksum = compute_file_sha256(downloaded_path)
-                lines.append("%s  %s" % (checksum, filename))
-                with open(checksum_file, 'wt') as out:
-                    for line in lines:
-                        out.write(line + "\n")
-                self.filename2checksum[filename] = checksum
-                log("Added checksum for %s to %s: %s", filename, checksum_file, checksum)
-                return checksum
-
-            fatal("No expected checksum provided for {}".format(filename))
-        return self.filename2checksum[filename]
-
-    def ensure_file_downloaded(self, url: str, path: str) -> None:
-        log(f"Ensuring {url} is downloaded to path {path}")
-        file_name = os.path.basename(path)
-
-        mkdir_if_missing(self.tp_download_dir)
-
-        if os.path.exists(path):
-            # We check the filename against our checksum map only if the file exists. This is done
-            # so that we would still download the file even if we don't know the checksum, making it
-            # easier to add new third-party dependencies.
-            expected_checksum = self.get_expected_checksum(file_name, downloaded_path=path)
-            if self.verify_checksum(path, expected_checksum):
-                log("No need to re-download %s: checksum already correct", file_name)
-                return
-            log("File %s already exists but has wrong checksum, removing", path)
-            remove_path(path)
-
-        log("Fetching %s from %s", file_name, url)
-        sleep_time_sec = INITIAL_DOWNLOAD_RETRY_SLEEP_TIME_SEC
-        for attempt_index in range(1, MAX_FETCH_ATTEMPTS + 1):
-            try:
-                subprocess.check_call([self.curl_path, '-o', path, '--location', url])
-                break
-            except subprocess.CalledProcessError as ex:
-                log("Error downloading %s (attempt %d): %s",
-                    self.curl_path, attempt_index, str(ex))
-                if attempt_index == MAX_FETCH_ATTEMPTS:
-                    log("Giving up after %d attempts", MAX_FETCH_ATTEMPTS)
-                    raise ex
-                log("Will retry after %.1f seconds", sleep_time_sec)
-                time.sleep(sleep_time_sec)
-                sleep_time_sec += DOWNLOAD_RETRY_SLEEP_INCREASE_SEC
-
-        if not os.path.exists(path):
-            fatal("Downloaded '%s' but but unable to find '%s'", url, path)
-        expected_checksum = self.get_expected_checksum(file_name, downloaded_path=path)
-        if not self.verify_checksum(path, expected_checksum):
-            fatal("File '%s' has wrong checksum after downloading from '%s'. "
-                  "Has %s, but expected: %s",
-                  path, url, compute_file_sha256(path), expected_checksum)
-
-    def verify_checksum(self, file_name: str, expected_checksum: str) -> bool:
-        real_checksum = hashsum_file(hashlib.sha256(), file_name)
-        return real_checksum == expected_checksum
-
-    def extract_archive(
-            self,
-            archive_file_name: str,
-            out_dir: str,
-            out_name: Optional[str] = None) -> None:
-        """
-        Extract the given archive into a subdirectory of out_dir, optionally renaming it to
-        the specified name out_name. The archive is expected to contain exactly one directory.
-        If out_name is not specified, the name of the directory inside the archive becomes
-        the name of the destination directory.
-
-        out_dir is the parent directory that should contain the extracted directory when the
-        function returns.
-        """
-
-        def dest_dir_already_exists(full_out_path: str) -> bool:
-            if os.path.exists(full_out_path):
-                log("Directory already exists: %s, skipping extracting %s" % (
-                        full_out_path, archive_file_name))
-                return True
-            return False
-
-        full_out_path = None
-        if out_name:
-            full_out_path = os.path.join(out_dir, out_name)
-            if dest_dir_already_exists(full_out_path):
-                return
-
-        # Extract the archive into a temporary directory.
-        tmp_out_dir = os.path.join(
-            out_dir, 'tmp-extract-%s-%s-%d' % (
-                os.path.basename(archive_file_name),
-                datetime.now().strftime('%Y-%m-%dT%H_%M_%S'),  # Current second-level timestamp.
-                random.randint(10 ** 8, 10 ** 9 - 1)))  # A random 9-digit integer.
-        if os.path.exists(tmp_out_dir):
-            raise IOError("Just-generated unique directory name already exists: %s" % tmp_out_dir)
-        os.makedirs(tmp_out_dir)
-
-        archive_extension = None
-        for ext in ARCHIVE_TYPES:
-            if archive_file_name.endswith(ext):
-                archive_extension = ext
-                break
-        if not archive_extension:
-            fatal("Unknown archive type for: {}".format(archive_file_name))
-        assert archive_extension is not None
-
-        try:
-            with PushDir(tmp_out_dir):
-                cmd = ARCHIVE_TYPES[archive_extension].format(archive_file_name)
-                log("Extracting %s in temporary directory %s", cmd, tmp_out_dir)
-                subprocess.check_call(cmd, shell=True)
-                extracted_subdirs = [
-                    subdir_name for subdir_name in os.listdir(tmp_out_dir)
-                    if not subdir_name.startswith('.')
-                ]
-                if len(extracted_subdirs) != 1:
-                    raise IOError(
-                        "Expected the extracted archive %s to contain exactly one "
-                        "subdirectory and no files, found: %s" % (
-                            archive_file_name, extracted_subdirs))
-                extracted_subdir_basename = extracted_subdirs[0]
-                extracted_subdir_path = os.path.join(tmp_out_dir, extracted_subdir_basename)
-                if not os.path.isdir(extracted_subdir_path):
-                    raise IOError(
-                        "This is a file, expected it to be a directory: %s" %
-                        extracted_subdir_path)
-
-                if not full_out_path:
-                    full_out_path = os.path.join(out_dir, extracted_subdir_basename)
-                    if dest_dir_already_exists(full_out_path):
-                        return
-
-                log("Moving %s to %s", extracted_subdir_path, full_out_path)
-                shutil.move(extracted_subdir_path, full_out_path)
-        finally:
-            log("Removing temporary directory: %s", tmp_out_dir)
-            shutil.rmtree(tmp_out_dir)
 
     def prepare_out_dirs(self) -> None:
         build_types: List[str] = list(BUILD_TYPES)
@@ -1191,7 +959,11 @@ class Builder(BuilderInterface):
         colored_log(YELLOW_COLOR, "Building %s (%s)", dep.name, self.build_type)
         colored_log(YELLOW_COLOR, SEPARATOR)
 
-        self.download_dependency(dep)
+        self.download_manager.download_dependency(
+            dep=dep,
+            src_path=self.get_source_path(dep),
+            archive_path=self.get_archive_path(dep))
+
         if self.args.download_extract_only:
             log("Skipping build of dependency %s, build type %s, --download-extract-only is "
                 "specified.", dep.name, self.build_type)
